todo edited

+++ app/src/main/java/com/example/todo/MainActivity.kt
@@ -1,12 +1,21 @@
 package com.example.todo
 
 import android.content.Intent
+import android.graphics.Color
+import android.graphics.Color.parseColor
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.os.PersistableBundle
+import android.view.View
+import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.core.content.ContentProviderCompat.requireContext
 import androidx.recyclerview.widget.LinearLayoutManager
 import kotlinx.android.synthetic.main.activity_create_card.*
+import kotlinx.android.synthetic.main.activity_create_card.create_title
 import kotlinx.android.synthetic.main.activity_main.*
+import kotlinx.android.synthetic.main.fragment_detail.*
+import kotlinx.android.synthetic.main.view.*
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -17,23 +26,33 @@ class MainActivity : AppCompatActivity() {
             var dialog = detailFragment()
             dialog.show(supportFragmentManager,dialog.tag)
         }
+
         deleteAll.setOnClickListener {
-            DataObject.deleteAll()
-            setRecycler()
+            if (DataObject.getAllData().isNotEmpty()) {
+                val builder = AlertDialog.Builder(this)
+                // Create the AlertDialog
+                //set title for alert dialog
+                builder.setTitle(R.string.dialogTitle)
+                //set message for alert dialog
+                builder.setMessage(R.string.dialogMessage)
+                builder.setIcon(android.R.drawable.ic_dialog_alert)
+
+                //performing positive action
+                builder.setPositiveButton("Yes") { dialogInterface, which ->
+                    DataObject.deleteAll()
+                    setRecycler()
+                }
+                //performing cancel action
+                builder.setNegativeButton("No") { dialogInterface, which ->
+                }
+                val alertDialog: AlertDialog = builder.create()
+
+                // Set other dialog properties
+                alertDialog.setCancelable(false)
+                alertDialog.show()
+            }
         }
         setRecycler()
-    }
-
-
-
-    private fun save() {
-        if(create_title.text.toString().trim{it<=' '}.isNotEmpty()
-            &&create_priority.text.toString().trim{it<=' '}.isNotEmpty()){
-            var title=create_title.getText().toString()
-            var priority=create_priority.getText().toString()
-            DataObject.setData(title, priority)}
-        var intent = Intent(this,MainActivity::class.java)
-        startActivity(intent)
     }
